requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.29
alembic>=1.13.1
textblob>=0.18.0
celery>=5.3.6
redis>=5.0.3
xrpl-py>=3.0.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.0
requests>=2.31.0
pytest>=8.0.0
httpx>=0.27.0
openai>=1.16.0
websockets>=12.0
nltk>=3.8.1
